Yixin modified Makefile, get rid of all warnings; uploaded two function.cpp files
