+++ .gitignore
@@ -1,4 +1,6 @@
 prototypes.o
+read_in.o
+mesh_gen_o.o
 main.o
 myprogram
 login.txt
